Move tests from tests/pytestto tests
