+++ pytest.ini
@@ -3,4 +3,4 @@ addopts =
     --cov-config=.coveragerc
     --cov=./
 testpaths =
-    test/pytest
+    test/
