+++ test/TestHardwareObjectBase.py
