+++ test/conftest.py
@@ -31,7 +31,7 @@ monkey.patch_all(thread=False)
 
 
 TESTS_DIR = os.path.abspath(os.path.dirname(__file__))
-ROOT_DIR = os.path.abspath(os.path.join(os.path.dirname(__file__), "../.."))
+ROOT_DIR = os.path.abspath(os.path.join(os.path.dirname(__file__), ".."))
 
 sys.path.insert(0, ROOT_DIR)
