+++ test/data/epics_channels.yaml
