+++ test/data/exporter_channels.yaml
