+++ test/data/exporter_commands_channels.yaml
