+++ test/data/tango_commands_channels.yaml
