+++ test/test_base_hardware_objects.py
