+++ test/test_detector.py
@@ -28,7 +28,7 @@ import math
 
 import pytest
 
-from test.pytest import TestHardwareObjectBase
+from test import TestHardwareObjectBase
 
 __copyright__ = """ Copyright © 2016 - 2020 by MXCuBE Collaboration """
 __license__ = "LGPLv3+"
