+++ test/test_resolution.py
@@ -22,7 +22,7 @@ hardware objects
 
 import pytest
 
-from test.pytest import TestAbstractMotorBase
+from test import TestAbstractMotorBase
 
 __copyright__ = """ Copyright © 2016 - 2022 by MXCuBE Collaboration """
 __license__ = "LGPLv3+"
