+++ test/test_transmission.py
@@ -25,7 +25,7 @@ __license__ = "LGPLv3+"
 
 import pytest
 
-from test.pytest import TestAbstractActuatorBase
+from test import TestAbstractActuatorBase
 
 
 @pytest.fixture
